Move to_cached_record_batch_stream to cache crate
